pyproject.toml
[tool.poetry]
name = "discord-bouncer"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
stripe = "^9.4.0"
flask = "^3.0.3"

[tool.poetry.scripts]
bouncer = "discord_bouncer.discord_bouncer:start_bouncer"
stripe = "discord_bouncer.checkout_session:handle_checkout_session"
flask = "discord_bouncer.app:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
